refactor: move std conversions into common module
