+++ src/conversions/mod.rs
@@ -1,13 +1,11 @@
 //! This module contains conversions between various Rust object and their representation in Python.
 
 pub mod anyhow;
-mod array;
 pub mod chrono;
 pub mod eyre;
 pub mod hashbrown;
 pub mod indexmap;
 pub mod num_bigint;
 pub mod num_complex;
-mod osstr;
-mod path;
 pub mod serde;
+mod std;
