+++ src/conversions/std/mod.rs
@@ -1,0 +1,9 @@
+mod array;
+mod map;
+mod num;
+mod osstr;
+mod path;
+mod set;
+mod slice;
+mod string;
+mod vec;
